try fix test
