+++ tests/integration/test_text_log_level/configs/config.d/text_log.xml
