+++ tests/queries/0_stateless/01473_event_time_microseconds.reference
@@ -6,3 +6,5 @@ ok
 ok
 01473_query_thread_log_table_event_start_time_microseconds_test
 ok
+01473_text_log_table_event_start_time_microseconds_test
+ok
