+++ tests/queries/skip_list.json
@@ -17,6 +17,7 @@
         "00151_replace_partition_with_different_granularity",
         "00157_cache_dictionary",
         "01193_metadata_loading",
+        "01473_event_time_microseconds",
         "01474_executable_dictionary" /// informational stderr from sanitizer at start
     ],
     "address-sanitizer": [
@@ -25,6 +26,7 @@
         "memory_profiler",
         "odbc_roundtrip",
         "01103_check_cpu_instructions_at_startup",
+        "01473_event_time_microseconds",
         "01193_metadata_loading"
     ],
     "ub-sanitizer": [
@@ -33,6 +35,7 @@
         "memory_profiler",
         "01103_check_cpu_instructions_at_startup",
         "00900_orc_load",
+        "01473_event_time_microseconds",
         "01193_metadata_loading"
     ],
     "memory-sanitizer": [
@@ -43,6 +46,7 @@
         "01086_odbc_roundtrip", /// can't pass because odbc libraries are not instrumented
         "00877_memory_limit_for_new_delete", /// memory limits don't work correctly under msan because it replaces malloc/free
         "01114_mysql_database_engine_segfault", /// it fails in _nss_files_parse_servent while using NSS from GLibc to authenticate (need to get rid of it)
+        "01473_event_time_microseconds",
         "01193_metadata_loading"
     ],
     "debug-build": [
@@ -57,6 +61,7 @@
         "01037_polygon_dicts_",
         "hyperscan",
         "01193_metadata_loading",
+        "01473_event_time_microseconds",
         "01396_inactive_replica_cleanup_nodes"
     ],
     "unbundled-build": [
